Fixed some tests run locally (#1187)
Always tokenize as batch to return `attention_mask` to preserve the
error:
```
The attention mask is not set and cannot be inferred from input because pad token is same as eos token. As a consequence, you may observe unexpected behavior. Please pass your input's `attention_mask` to obtain reliable results.
```