Update VLM example code in README (#1466)
Add `pipe.start_chat()` to VLM example. Without this, inference with
several models results in empty outputs.

This can be removed if this will be the default for VLM models, but at
the moment, the most basic example should work with supported models.

Also changed printing the VLMDecodedResults to getting the generated
text and printing that (see comment from Ilya).