[tokenizer] Fix setting max_legth and special tokens flags (#1860)
- `MakePaddingSatateful` was skipping left padding graphs during match
because it was checking for `Subtract` on the wrong `1st input`, while
it should've do this on `0th input`. We did miss that, because models
list in test by default was testing only right padding - now added
explicitly check left padding as well.
- There was an issue with `add_special_tokens`, Even when we called
`decode` it was rewriting `m_add_special_tokens` values as if we called
`set_state`, therefore if we subsequently requested `encode` even with
different params Tokenizer wrongly assumed that InferReuest already has
that param in state and did not modify it and we end up having wrong
output, e.g. below on the second call there should been 1 at the
begining. Fixed this issue as well.


![image](https://github.com/user-attachments/assets/79b3b27a-238a-4440-8ec6-4a85fd45e59b)

CVS-163154