Expose StructuredOutputConfig validation method (#2576)
Currently, as we pass JSON schema to structured output config as the
source of our grammar, it gets validated as grammar gets created and if
we pass bad schema there is an error.

Grammar creation happens in Sampler/LogitProcessor which means we learn
about bad schema when request is already in the pipeline and if the
schema is bad, the pipeline is broken.

This PR exposes additional `validate` method on StructuredOutputConfig
level that attempts to create grammar, so it's possible to check if
grammar is not malformed before the request is scheduled for processing.