Fix MiniCPM-o-2_6 (#2847)
## Description
WWB uses `_OVMiniCPMVForCausalLM.preprocess_inputs()` for
`optimum-intel` and `--hf`. That function fails to apply chat_template
for `MiniCPM-o-2_6` and `MiniCPM-V-2_6`. A fix is required for
`optimum-intel`. It will take time to validate both of the models with
that fix. This PR addresses other identified issues. The original
implementation applies chat_template:
https://huggingface.co/openbmb/MiniCPM-o-2_6/blob/main/modeling_minicpmo.py#L959
but it's not used in WWB.
WWB results for `--weights-format fp16` and unpublished fix for
`optimum-intel` chat_template:
1. --hf vs optimum-intel : 0.986911
2. --hf vs --genai: 0.935212
3. optimum-intel vs --genai 0.942425

Docs:
https://wovchena.github.io/openvino.genai-public/docs/supported-models/#visual-language-models-vlms

That should be enough for this release but follow up PRs are needed. The
ideal solution includes:
1. Fix optimum-intel chat_template application
2. Request WWB not to use optimum-intel with --hf
3. Align --hf and optimum-intel (likely to be never completed because
the results are close: 0.986911)
4. Extend GenAI pre-commit tests
5. Copy image resize implementation from PIL to GenAI
6. Check if it's possible to align `sin()` and `cos()` with `numpy` with
hardcoding commit `sin()` and `cos()` values

CVS-170987

## Checklist:
- [ ] Tests have been updated or added to cover the new code <!--- If
the change isn't maintenance related, update the tests at
https://github.com/openvinotoolkit/openvino.genai/tree/master/tests or
explain in the description why the tests don't need an update. -->
- [ ] This patch fully addresses the ticket. <!--- If follow-up pull
requests are needed, specify in description. -->
- [x] I have made corresponding changes to the documentation