[GHA] Fix samples dependabot duplicated PRs (#3031)
<!-- Keep your pull requests (PRs) as atomic as possible. That increases
the likelihood that an individual PR won't be stuck because of adjacent
problems, merge conflicts, or code review.
Your merged PR is going to appear in the automatically generated release
notes on GitHub. So the clearer the title the better. -->
## Description
Duplicated PR:
https://github.com/openvinotoolkit/openvino.genai/pull/3014
https://github.com/openvinotoolkit/openvino.genai/pull/3015
There is some inconsistency how `/samples` and `/tests` scanned. This is
a fix attempt. Group directories and prefix samples with `.`