[NPU] Add `batch_size` support for embedding model (#2986)
<!-- Keep your pull requests (PRs) as atomic as possible. That increases
the likelihood that an individual PR won't be stuck because of adjacent
problems, merge conflicts, or code review.
Your merged PR is going to appear in the automatically generated release
notes on GitHub. So the clearer the title the better. -->
## Description
<!-- Please include a summary of the change. Also include relevant
motivation and context. -->
The model `qwen3-embedding-0.6B` failed on `wwb` test on NPU due to
dynamic `batch size`. This PR adds `batch_size` option support for this
model in `llm-benchmark` and `wwb`.

<!-- Jira ticket number (e.g., 123). Delete if there's no ticket. -->
[CVS-176378](https://jira.devtools.intel.com/browse/CVS-176378)


## Checklist:
- [ ] Tests have been updated or added to cover the new code. <!-- If
the change isn't maintenance related, update the tests at
https://github.com/openvinotoolkit/openvino.genai/tree/master/tests or
explain in the description why the tests don't need an update. -->
- [ ] This patch fully addresses the ticket. <!--- If follow-up pull
requests are needed, specify in description. -->
- [ ] I have made corresponding changes to the documentation. <!-- Run
github.com/\<username>/openvino.genai/actions/workflows/deploy_gh_pages.yml
on your fork with your branch as a parameter to deploy a test version
with the updated content. Replace this comment with the link to the
built docs. -->

---------

Co-authored-by: Sofya Balandina <sofya.balandina@intel.com>