[JS] Bump js package up to 2026.0.0 and update dependecies (#3090)
## Description
- Bump JS package up to 2026.0.0
- Update JS dependencies

<!-- Jira ticket number (e.g., 123). Delete if there's no ticket. -->
CVS-177515

## Checklist:
- [ ] Tests have been updated or added to cover the new code. <!-- If
the change isn't maintenance related, update the tests at
https://github.com/openvinotoolkit/openvino.genai/tree/master/tests or
explain in the description why the tests don't need an update. -->
- [x] This patch fully addresses the ticket. <!--- If follow-up pull
requests are needed, specify in description. -->
- [ ] I have made corresponding changes to the documentation. <!-- Run
github.com/\<username>/openvino.genai/actions/workflows/deploy_gh_pages.yml
on your fork with your branch as a parameter to deploy a test version
with the updated content. Replace this comment with the link to the
built docs. -->

---------

Signed-off-by: Kirill Suvorov <kirill.suvorov@intel.com>