[llm_bench] Add batch considering in throughput and 2nd token latency for LLMPipeline (#3108)
<!-- Keep your pull requests (PRs) as atomic as possible. That increases
the likelihood that an individual PR won't be stuck because of adjacent
problems, merge conflicts, or code review.
Your merged PR is going to appear in the automatically generated release
notes on GitHub. So the clearer the title the better. -->
## Description
It was found that llm_bench shows very low throughput for speculative
decode. This happens because llm_bench calculates the latency for tokens
other than the first based on the `PerfMetrics`
`raw_metrics.m_new_token_times`, which do not take into account the
batch. `raw_metrics.m_durations` are also calculated based on
`raw_metrics.m_new_token_times`, but the time is divided into batches,
which significantly affects the calculations for speculative decoding.

<!-- Jira ticket number (e.g., 123). Delete if there's no ticket. -->
CVS-174513

<!-- Remove if not applicable -->
Fixes #(issue)

## Checklist:
- [ ] Tests have been updated or added to cover the new code. <!-- If
the change isn't maintenance related, update the tests at
https://github.com/openvinotoolkit/openvino.genai/tree/master/tests or
explain in the description why the tests don't need an update. -->
- [ ] This patch fully addresses the ticket. <!--- If follow-up pull
requests are needed, specify in description. -->
- [ ] I have made corresponding changes to the documentation. <!-- Run
github.com/\<username>/openvino.genai/actions/workflows/deploy_gh_pages.yml
on your fork with your branch as a parameter to deploy a test version
with the updated content. Replace this comment with the link to the
built docs. -->