[JS] VLMPipeline implementation (#3112)
<!-- Keep your pull requests (PRs) as atomic as possible. That increases
the likelihood that an individual PR won't be stuck because of adjacent
problems, merge conflicts, or code review.
Your merged PR is going to appear in the automatically generated release
notes on GitHub. So the clearer the title the better. -->
## Description
This PR introduces comprehensive support for the JavaScript bindings in
the VLMPipeline.

VLMPipeline is implemented similarly to LLMPipeline, but with some
improvements:
- Error handling in `vLmPerformInferenceThread` was modified for a much
better user experience. It no longer throws fatal errors. Now, the
thread uses a callback to reject the promise and throw a catchable
exception to the user. If that is not possible, an error message is
logged.
- Streamer and callback were separated to create clearer code. These
functions are used in different cases and have different semantics. A
callback is always used to resolve a promise with the final result or
reject it with an error. The streamer is only used if it is provided by
the user to get chunks.
- A wrapper property, `is_initializing`, was added to prevent the
pipeline from being loaded incorrectly twice.
- A wrapper property, `is_generating`, was added to prevent the
incorrect attempt to call generate in parallel. This leads to an
InferRequest exception, which may confuse the user.
- VLMPerfMetrics should extend PerfMetrics. This was refactored to avoid
code duplication. The Node addon API doesn't support inheritance from
the box, so a
[workaround](https://mmomtchev.medium.com/c-class-inheritance-with-node-api-and-node-addon-api-c180334d9902)
was used to resolve this issue.
- The optional parameters for generate and stream were grouped into the
options argument. This hook allows the use of named arguments and
simplifies calls.

<!-- Jira ticket number (e.g., 123). Delete if there's no ticket. -->
CVS-172789

## Checklist:
- [x] Tests have been updated or added to cover the new code. <!-- If
the change isn't maintenance related, update the tests at
https://github.com/openvinotoolkit/openvino.genai/tree/master/tests or
explain in the description why the tests don't need an update. -->
- [x] This patch fully addresses the ticket. <!--- If follow-up pull
requests are needed, specify in description. -->
- [x] I have made corresponding changes to the documentation.
https://retribution98.github.io/openvino.genai/

---------

Signed-off-by: Kirill Suvorov <kirill.suvorov@intel.com>
Co-authored-by: Yaroslav Tarkan <yaroslav.tarkan@intel.com>